topology: ABI - Add the types for BE DAI
Define the type and ABI struct for BE (Back End) DAIs. Add the number of
BE DAIs to manifest, and add some reserved fields for future extensions.

Pump the version number to 5.

Topology kernel driver will check size of ABI objects to detect version
mismatch between user space and kernel.

Signed-off-by: Guneshwor Singh <guneshwor.o.singh@intel.com>
Signed-off-by: Mengdong Lin <mengdong.lin@linux.intel.com>