topology: Add support for included files
Users can include a list of files by SectionInclude in the text config
file. This allows user to define common info in separate files (e.g.
vendor tokens, tuples) and share them for different platforms, by
including them. This can save the total size of files. Users can also
specifiy a list of directories relative to "usr/share/alsa/topology"
to search the included files.

The topology library will search and open an included file in the
following order of priority:
1. directly open the file by its name;
2. search for the file name in the directory containing current file;
3. search for the file name in "usr/share/alsa/topology";
4. search for the file name in user specified subdirectories under
   "usr/share/alsa/topology".

The order of the included files need not to be same as their dependencies,
because the toplogy library will load all of them before parsing their
dependencies.

Signed-off-by: Mengdong Lin <mengdong.lin@linux.intel.com>